Upload mydatabase.mwb
A .mwb file is a data model file used by MySQL Workbench, a graphical tool for managing and designing databases. It stores the structure of a database, including tables, columns, and relationships. You can open a .mwb file in MySQL Workbench using the "Open Model" option in the File menu.